Fixed `--graph`/`--no-graph` and `--patch` OOM.
Currently using `--graph` and `--patch` together may run out of memory for large repos.
Also `--no-graph` undoes `--graph` if the former is given after the latter.
However, if `revs->graph` is initialized at least once, the arguments parser function consider it somewhere in invoked code (where?) in the parsing loop.
This causes the OOM.
I have no idea how the `revs` object is processed (and I don't really care) within the parsing loop, so this is why I implemented a simple `if --graph is not overridden with --no-graph` check.
Yeah, this check is O(n) causing the outer loop to grow exponentially (depends on the argc value though).
No idea why the arguments parser does not merely sets flags first, and then, and only after that, does all required initialization.
I don't care the hell it is implemented.
This stupid patch just works.